feat: Add Canada to countries.json and update login script source
This commit adds Canada to the list of countries in the `countries.json` file. It also updates the source of the login script in the `login.html` file to `/api/script/login.js`. These changes enhance the functionality and data of the application.

Note: This commit message follows the conventional format of "feat: <description>".